Fixes URLs with query strings and hashes in them The previous fix did not test a corner case where the user passes extensions allowed in the options. When there are no otpions passed for extensions allowed the extension check is skipped and thus the file is processed appropriately.
I added virtual copies of the tests so feel free to
point me to a better way of doing that. I just don't
know what a better way to do it is without overwriting
the original tests.

The fix I made was stripping query strings and hashes
from the regex results before the extension check and
then passing in that stripped version into the check
instead of the actual result which has the hash and/or
query string.